Start commits for video:Refactoring Massive View Controllers Into Multiple Tiny MVCs -Extract UIImage test helper to a separate file
