+++ EssentialFeedJorgeiOSTests/FeedViewControllerTests.swift
@@ -470,23 +470,6 @@ private extension UIRefreshControl {
 }
 
 
-// MARK: - Extension. UIImage
-private extension UIImage {
-    static func make(withColor color: UIColor) -> UIImage {
-        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
-        let format = UIGraphicsImageRendererFormat()
-        format.scale = 1
-        
-        let imgRenderer = UIGraphicsImageRenderer(size: rect.size, format: format).image { rendererContext in
-            color.setFill()
-            rendererContext.fill(rect)
-        }
-        
-        return imgRenderer
-    }
-}
-
-
 // MARK: - Extension. UIButton
 private extension UIButton {
     func simulateTap() {
