+++ EssentialFeedJorgeiOSTests/Helpers/UIImage+TestHelpers.swift
@@ -1,0 +1,26 @@
+//
+//  UIImage+TestHelpers.swift
+//  EssentialFeedJorgeiOSTests
+//
+//  Created by jfuentescasillas on 15/09/23.
+//
+
+
+import UIKit
+
+
+// MARK: - Extension. UIImage
+extension UIImage {
+    static func make(withColor color: UIColor) -> UIImage {
+        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
+        let format = UIGraphicsImageRendererFormat()
+        format.scale = 1
+        
+        let imgRenderer = UIGraphicsImageRenderer(size: rect.size, format: format).image { rendererContext in
+            color.setFill()
+            rendererContext.fill(rect)
+        }
+        
+        return imgRenderer
+    }
+}
